[Cherry-pick][Bugfix] Align swiglu limit attr type (#16128)
### What this PR does / why we need it:
Cherry-pick of #15933 to main.

Cast `limited` to float before passing it into the L0 OP_ATTR path for
GroupedMatmulSwigluQuant, matching the OpDef Float attr and
static-kernel dump expectations. Also read the attr as float in tiling
with a double fallback.

Backport of #15933 (merged into releases/v0.26.0rc).

- vLLM main:
https://github.com/vllm-project/vllm/commit/b2f685834a6456197e7033966fdef52a23f1abcd

+++ csrc/gmm/grouped_matmul_swiglu_quant/op_host/op_api/grouped_matmul_swiglu_quant.cpp
@@ -20,7 +20,7 @@ OP_TYPE_REGISTER(GroupedMatmulSwigluQuant);
 
 const std::tuple<aclTensor *, aclTensor *>
 GroupedMatmulSwigluQuant(const aclTensor *x, const aclTensor *weight, const aclTensor *perChannelScale,
-                         const aclTensor *perTokenScale, const aclTensor *groupList, double limited,
+                         const aclTensor *perTokenScale, const aclTensor *groupList, float limited,
                          const aclTensor *weightAssistanceMatrix, bool isEnableWeightAssistanceMatrix, int dequantMode,
                          aclOpExecutor *executor)
 {
